add MPI support to libdeltafs.   if MPI is enabled in the build, then anything that uses the deltafs library target also gets an MPI dependency.
notes: we are currently using FindMPI.cmake module that comes with
cmake.  This module probes the current system for an installed MPI
and uses its compiler wrapper programs (mpicc, mpicxx) to determine
the location of the MPI include files and libs.  It returns this
info in variables MPI_<lang>_* (where * is COMPILE_FLAGS, INCLUDE_PATH,
LINK_FLAGS, and LIBRARIES).  It also returns the compiler wrapper
in MPI_<lang>_COMPILER.  This gives you two options: either add
the MPI include path, and libs to your build and use the normal
compiler, or override the compiler variables with the wrapper and
use that.  Currently we are using the first option (add the MPI
depends directly to libdeltafs and don't use the MPI compiler wrappers).
Hopefully that will work, otherwise we'll have to fall back to the
wrappers.

Note that the FindMPI.cmake module does not create any targets,
so we add the MPI compile/linkage information to the deltafs target.