add compaction manager
a compaction manager manages generic compaction state for objects
that serially write their data through compaction to backing store.
we ensure that compactions complete in sequence number order.
we use our parent's mutex lock/cv for locking and sync.

the idea is to embed this in an object you want to manage.

common high level APIs:
  - ScheduleCompaction() -- schedule a background compaction.  bgnd
    thread will call the given callback when it is time to compact.

  - AquireWriteToken() -- called in callback when we want to write.
    will block if not our turn to write yet.

  - WaitForAll() -- wait for all pending compactions to complete

  - Finish() -- we are done, finish all compactions