add range_writer backend for CARP
This adds DELTAFS_PLFSDIR_RANGEDB: a new io_engine backend for CARP
range queires.  There is also a new api call deltafs_plfsdir_range_update()
for updating ranges.   This code is only active if the io_engine is
requested, so it should not have any impact on other modes of
operation.