have Range and ObservedRange (the later tracks key metadata)
A Range has a min (inclusive) and max (exclusive) and can be
Set() or Extended().  You can test to see if it IsSet() or
IsValid(), and you can test to see if a point is Inside()
the Range.

An ObservedRange does what range does, but also has an Observe(value)
function that tracks the smallest and largest values inserted in the
ObservedRange.  It also tracks the number of items observed and how
many of those items were out of bounds.

ordered builder now only needs one ObservedRange to track expected
range and observed values.  partition manifest writer can also use
ObservedRange to track overall metadata in an manifest_obrange_ object.

add unit test for range.h