resolve assert during initial range updates
range.h: allow Set() to set ranges to the default (reset) values.
This happens at startup in range_writer.  the initial set of blocks
are all reset.  When we update ranges, we slowly push the default
values out of the bactive_[] stack as we update.  also make sure
zero-width bins are allowed in InValid().

ordered_builder.h: get rid of "rmax >= rmin" assert since this
can happen at init (see above) and it is also checked below
in the obrange_.Set() call, so we don't need to do a redundant
check.

range_writer: add a check to UpdateBounds to see if it is trying
to update bounds to a reset state (not expected to happen, so give
a shout if it does).